Reduce memory footprint in image enhancement.
When updating to an album with 83 new images, the previous code would
stop working after a few were processed. There was no specific error
message; the process just stopped working. This change sacrifices some
speed for lowering the amount of memory consumed, allowing for all 83
photos to be downloaded and processed.