Added resuming interrupted image processing
If a series of images are downloaded, they need to be processed. If that
processing is interrupted due to something like the program crashing,
some were left unprocessed. This patch introduces a flag to indicate
that processing didn't finish.

It is not an ideal solution, as it can't distinguish between those
images that have been processed and those that haven't, so some may end
up having the adjustments applied more than once. I'll work on that
next.