Improved unprocessed image handling
The previous approach was a global flag that could not distinguish
between images that had been processed and those that had not; it only
was able to say that the overall processing did not finish.

This downloads images to a separate directory, and then moves the
processed image to the images/ directory. If processing is interrupted,
only the images remaining in the download directory are processed next
time the program runs.