Incorporated etcd for server communication
This commit removes the 'check_host' functionality that essentially
polled the server on a regular basis in case there were any changes that
needed to be applied. Replacing it is code that watches an etcd server
for changes to keys for this viewer. When a key is changed, the code
carries out the changes for the new value.