Changed the etcd key format
Previously the key was of the form "/photoviewer:{pkid}", and the value
was a dict with "topic" and "data" as the keys. One problem with that is
that any time the server wrote anything, it overwrote what was there. I
want to have the viewer be able to query its state, and for that the key
format needed to change. The new format is "/{pkid}:{action}", and the
value is a JSON-encoded string of the data.